+++ Services/DataService.cs
@@ -24,6 +24,16 @@ internal static class DataService
         Mining,
         Fishing
     }
+    public enum PlayerClass
+    {
+        None,
+        BloodKnight,
+        DemonHunter,
+        VampireLord,
+        ShadowBlade,
+        ArcaneSorcerer,
+        DeathMage
+    }
     public enum BloodType
     {
         Worker,
@@ -85,7 +95,7 @@ internal static class DataService
         { WeaponStatType.PrimaryAttackSpeed, "PAS" },
         { WeaponStatType.PhysicalLifeLeech, "PLL" },
         { WeaponStatType.SpellLifeLeech, "SLL" },
-        { WeaponStatType.PrimaryLifeLeech, "PLL" },
+        { WeaponStatType.PrimaryLifeLeech, "PAL" },
         { WeaponStatType.PhysicalPower, "PP" },
         { WeaponStatType.SpellPower, "SP" },
         { WeaponStatType.PhysicalCriticalStrikeChance, "PCC" },
@@ -101,7 +111,7 @@ internal static class DataService
         { "PrimaryAttackSpeed", "PAS" },
         { "PhysicalLifeLeech", "PLL" },
         { "SpellLifeLeech", "SLL" },
-        { "PrimaryLifeLeech", "PLL" },
+        { "PrimaryLifeLeech", "PAL" },
         { "PhysicalPower", "PP" },
         { "SpellPower", "SP" },
         { "PhysicalCriticalStrikeChance", "PCC" },
@@ -209,38 +219,6 @@ internal static class DataService
         { Profession.Mining,        new Color(0.5f, 0.5f, 0.5f, 0.5f) },
         { Profession.Fishing,       new Color(0f, 0.5f, 0.7f, 0.5f) }
     };
-    public class ProfessionData(string enchantingProgress, string enchantingLevel, string alchemyProgress, string alchemyLevel,
-        string harvestingProgress, string harvestingLevel, string blacksmithingProgress, string blacksmithingLevel,
-        string tailoringProgress, string tailoringLevel, string woodcuttingProgress, string woodcuttingLevel, string miningProgress,
-        string miningLevel, string fishingProgress, string fishingLevel)
-    {
-        public float EnchantingProgress { get; set; } = float.Parse(enchantingProgress, CultureInfo.InvariantCulture) / 100f;
-        public int EnchantingLevel { get; set; } = int.Parse(enchantingLevel);
-        public float AlchemyProgress { get; set; } = float.Parse(alchemyProgress, CultureInfo.InvariantCulture) / 100f;
-        public int AlchemyLevel { get; set; } = int.Parse(alchemyLevel);
-        public float HarvestingProgress { get; set; } = float.Parse(harvestingProgress, CultureInfo.InvariantCulture) / 100f;
-        public int HarvestingLevel { get; set; } = int.Parse(harvestingLevel);
-        public float BlacksmithingProgress { get; set; } = float.Parse(blacksmithingProgress, CultureInfo.InvariantCulture) / 100f;
-        public int BlacksmithingLevel { get; set; } = int.Parse(blacksmithingLevel);
-        public float TailoringProgress { get; set; } = float.Parse(tailoringProgress, CultureInfo.InvariantCulture) / 100f;
-        public int TailoringLevel { get; set; } = int.Parse(tailoringLevel);
-        public float WoodcuttingProgress { get; set; } = float.Parse(woodcuttingProgress, CultureInfo.InvariantCulture) / 100f;
-        public int WoodcuttingLevel { get; set; } = int.Parse(woodcuttingLevel);
-        public float MiningProgress { get; set; } = float.Parse(miningProgress, CultureInfo.InvariantCulture) / 100f;
-        public int MiningLevel { get; set; } = int.Parse(miningLevel);
-        public float FishingProgress { get; set; } = float.Parse(fishingProgress, CultureInfo.InvariantCulture) / 100f;
-        public int FishingLevel { get; set; } = int.Parse(fishingLevel);
-    }
-    public enum PlayerClass
-    {
-        None,
-        BloodKnight,
-        DemonHunter,
-        VampireLord,
-        ShadowBlade,
-        ArcaneSorcerer,
-        DeathMage
-    }
 
     public static Dictionary<PlayerClass, (List<WeaponStatType> WeaponStats, List<BloodStatType> BloodStats)> _classStatSynergies = [];
 
@@ -248,28 +226,50 @@ internal static class DataService
     public static float _classStatMultiplier;
     public static bool _extraRecipes;
     public static PrefabGUID _primalCost;
+    public class ProfessionData(string enchantingProgress, string enchantingLevel, string alchemyProgress, string alchemyLevel,
+        string harvestingProgress, string harvestingLevel, string blacksmithingProgress, string blacksmithingLevel,
+        string tailoringProgress, string tailoringLevel, string woodcuttingProgress, string woodcuttingLevel, string miningProgress,
+        string miningLevel, string fishingProgress, string fishingLevel)
+    {
+        public float EnchantingProgress { get; set; } = float.Parse(enchantingProgress, CultureInfo.InvariantCulture) / 100f;
+        public int EnchantingLevel { get; set; } = int.Parse(enchantingLevel, CultureInfo.InvariantCulture);
+        public float AlchemyProgress { get; set; } = float.Parse(alchemyProgress, CultureInfo.InvariantCulture) / 100f;
+        public int AlchemyLevel { get; set; } = int.Parse(alchemyLevel, CultureInfo.InvariantCulture);
+        public float HarvestingProgress { get; set; } = float.Parse(harvestingProgress, CultureInfo.InvariantCulture) / 100f;
+        public int HarvestingLevel { get; set; } = int.Parse(harvestingLevel, CultureInfo.InvariantCulture);
+        public float BlacksmithingProgress { get; set; } = float.Parse(blacksmithingProgress, CultureInfo.InvariantCulture) / 100f;
+        public int BlacksmithingLevel { get; set; } = int.Parse(blacksmithingLevel, CultureInfo.InvariantCulture);
+        public float TailoringProgress { get; set; } = float.Parse(tailoringProgress, CultureInfo.InvariantCulture) / 100f;
+        public int TailoringLevel { get; set; } = int.Parse(tailoringLevel, CultureInfo.InvariantCulture);
+        public float WoodcuttingProgress { get; set; } = float.Parse(woodcuttingProgress, CultureInfo.InvariantCulture) / 100f;
+        public int WoodcuttingLevel { get; set; } = int.Parse(woodcuttingLevel, CultureInfo.InvariantCulture);
+        public float MiningProgress { get; set; } = float.Parse(miningProgress, CultureInfo.InvariantCulture) / 100f;
+        public int MiningLevel { get; set; } = int.Parse(miningLevel, CultureInfo.InvariantCulture);
+        public float FishingProgress { get; set; } = float.Parse(fishingProgress, CultureInfo.InvariantCulture) / 100f;
+        public int FishingLevel { get; set; } = int.Parse(fishingLevel, CultureInfo.InvariantCulture);
+    }
     public class ExperienceData(string percent, string level, string prestige, string playerClass)
     {
         public float Progress { get; set; } = float.Parse(percent, CultureInfo.InvariantCulture) / 100f;
-        public int Level { get; set; } = int.Parse(level);
-        public int Prestige { get; set; } = int.Parse(prestige);
-        public PlayerClass Class { get; set; } = (PlayerClass)int.Parse(playerClass);
+        public int Level { get; set; } = int.Parse(level, CultureInfo.InvariantCulture);
+        public int Prestige { get; set; } = int.Parse(prestige, CultureInfo.InvariantCulture);
+        public PlayerClass Class { get; set; } = (PlayerClass)int.Parse(playerClass, CultureInfo.InvariantCulture);
     }
     public class LegacyData(string percent, string level, string prestige, string legacyType, string bonusStats) : ExperienceData(percent, level, prestige, legacyType)
     {
-        public string LegacyType { get; set; } = ((BloodType)int.Parse(legacyType)).ToString();
-        public List<string> BonusStats { get; set; } = Enumerable.Range(0, bonusStats.Length / 2).Select(i => ((BloodStatType)int.Parse(bonusStats.Substring(i * 2, 2))).ToString()).ToList();
+        public string LegacyType { get; set; } = ((BloodType)int.Parse(legacyType, CultureInfo.InvariantCulture)).ToString();
+        public List<string> BonusStats { get; set; } = [..Enumerable.Range(0, bonusStats.Length / 2).Select(i => ((BloodStatType)int.Parse(bonusStats.Substring(i * 2, 2), CultureInfo.InvariantCulture)).ToString())];
     }
     public class ExpertiseData(string percent, string level, string prestige, string expertiseType, string bonusStats) : ExperienceData(percent, level, prestige, expertiseType)
     {
         public string ExpertiseType { get; set; } = ((WeaponType)int.Parse(expertiseType)).ToString();
-        public List<string> BonusStats { get; set; } = Enumerable.Range(0, bonusStats.Length / 2).Select(i => ((WeaponStatType)int.Parse(bonusStats.Substring(i * 2, 2))).ToString()).ToList();
+        public List<string> BonusStats { get; set; } = [.. Enumerable.Range(0, bonusStats.Length / 2).Select(i => ((WeaponStatType)int.Parse(bonusStats.Substring(i * 2, 2), CultureInfo.InvariantCulture)).ToString())];
     }
     public class QuestData(string type, string progress, string goal, string target, string isVBlood)
     {
-        public TargetType TargetType { get; set; } = (TargetType)int.Parse(type);
-        public int Progress { get; set; } = int.Parse(progress);
-        public int Goal { get; set; } = int.Parse(goal);
+        public TargetType TargetType { get; set; } = (TargetType)int.Parse(type, CultureInfo.InvariantCulture);
+        public int Progress { get; set; } = int.Parse(progress, CultureInfo.InvariantCulture);
+        public int Goal { get; set; } = int.Parse(goal, CultureInfo.InvariantCulture);
         public string Target { get; set; } = target;
         public bool IsVBlood { get; set; } = bool.Parse(isVBlood);
     }
@@ -277,13 +277,13 @@ internal static class DataService
     {
         public float Progress { get; set; } = float.Parse(percent, CultureInfo.InvariantCulture) / 100f;
         public int Level { get; set; } = int.TryParse(level, out int parsedLevel) && parsedLevel > 0 ? parsedLevel : 1;
-        public int Prestige { get; set; } = int.Parse(prestige);
+        public int Prestige { get; set; } = int.Parse(prestige, CultureInfo.InvariantCulture);
         public string FamiliarName { get; set; } = !string.IsNullOrEmpty(familiarName) ? familiarName : "Familiar";
-        public List<string> FamiliarStats { get; set; } = !string.IsNullOrEmpty(familiarStats) ? [..new List<string> { familiarStats[..4], familiarStats[4..7], familiarStats[7..] }.Select(stat => int.Parse(stat).ToString())] : ["", "", ""];
+        public List<string> FamiliarStats { get; set; } = !string.IsNullOrEmpty(familiarStats) ? [..new List<string> { familiarStats[..4], familiarStats[4..7], familiarStats[7..] }.Select(stat => int.Parse(stat, CultureInfo.InvariantCulture).ToString())] : ["", "", ""];
     }
     public class ShiftSpellData(string index)
     {
-        public int ShiftSpellIndex { get; set; } = int.Parse(index);
+        public int ShiftSpellIndex { get; set; } = int.Parse(index, CultureInfo.InvariantCulture);
     }
     public class ConfigDataV1_3
     {
@@ -315,14 +315,14 @@ internal static class DataService
             PrestigeStatMultiplier = float.Parse(prestigeMultiplier, CultureInfo.InvariantCulture);
             ClassStatMultiplier = float.Parse(statSynergyMultiplier, CultureInfo.InvariantCulture);
 
-            MaxPlayerLevel = int.Parse(maxPlayerLevel);
-            MaxLegacyLevel = int.Parse(maxLegacyLevel);
-            MaxExpertiseLevel = int.Parse(maxExpertiseLevel);
-            MaxFamiliarLevel = int.Parse(maxFamiliarLevel);
-            MaxProfessionLevel = int.Parse(maxProfessionLevel);
+            MaxPlayerLevel = int.Parse(maxPlayerLevel, CultureInfo.InvariantCulture);
+            MaxLegacyLevel = int.Parse(maxLegacyLevel, CultureInfo.InvariantCulture);
+            MaxExpertiseLevel = int.Parse(maxExpertiseLevel, CultureInfo.InvariantCulture);
+            MaxFamiliarLevel = int.Parse(maxFamiliarLevel, CultureInfo.InvariantCulture);
+            MaxProfessionLevel = int.Parse(maxProfessionLevel, CultureInfo.InvariantCulture);
 
             ExtraRecipes = bool.Parse(extraRecipes);
-            PrimalCost = int.Parse(primalCost);
+            PrimalCost = int.Parse(primalCost, CultureInfo.InvariantCulture);
 
             WeaponStatValues = weaponStatValues.Split(',')
             .Select((value, index) => new { Index = index + 1, Value = float.Parse(value, CultureInfo.InvariantCulture) })
@@ -337,70 +337,13 @@ internal static class DataService
             .Select((value, index) => new { Value = value, Index = index })
             .GroupBy(x => x.Index / 3)
             .ToDictionary(
-                g => (PlayerClass)int.Parse(g.ElementAt(0).Value),
+                g => (PlayerClass)int.Parse(g.ElementAt(0).Value, CultureInfo.InvariantCulture),
                 g => (
                     Enumerable.Range(0, g.ElementAt(1).Value.Length / 2)
-                        .Select(j => (WeaponStatType)int.Parse(g.ElementAt(1).Value.Substring(j * 2, 2)))
+                        .Select(j => (WeaponStatType)int.Parse(g.ElementAt(1).Value.Substring(j * 2, 2), CultureInfo.InvariantCulture))
                         .ToList(),
                     Enumerable.Range(0, g.ElementAt(2).Value.Length / 2)
-                        .Select(j => (BloodStatType)int.Parse(g.ElementAt(2).Value.Substring(j * 2, 2)))
-                        .ToList()
-                )
-            );
-        }
-    }
-    public class ConfigData
-    {
-        public float PrestigeStatMultiplier;
-
-        public float ClassStatMultiplier;
-
-        public int MaxPlayerLevel;
-
-        public int MaxLegacyLevel;
-
-        public int MaxExpertiseLevel;
-
-        public int MaxFamiliarLevel;
-
-        public int MaxProfessionLevel;
-
-        public Dictionary<WeaponStatType, float> WeaponStatValues;
-
-        public Dictionary<BloodStatType, float> BloodStatValues;
-
-        public Dictionary<PlayerClass, (List<WeaponStatType> WeaponStats, List<BloodStatType> BloodStats)> ClassStatSynergies;
-        public ConfigData(string prestigeMultiplier, string statSynergyMultiplier, string maxPlayerLevel, string maxLegacyLevel, string maxExpertiseLevel, string maxFamiliarLevel, string maxProfessionLevel, string weaponStatValues, string bloodStatValues, string classStatSynergies)
-        {
-            PrestigeStatMultiplier = float.Parse(prestigeMultiplier, CultureInfo.InvariantCulture);
-            ClassStatMultiplier = float.Parse(statSynergyMultiplier, CultureInfo.InvariantCulture);
-
-            MaxPlayerLevel = int.Parse(maxPlayerLevel);
-            MaxLegacyLevel = int.Parse(maxLegacyLevel);
-            MaxExpertiseLevel = int.Parse(maxExpertiseLevel);
-            MaxFamiliarLevel = int.Parse(maxFamiliarLevel);
-            MaxProfessionLevel = int.Parse(maxProfessionLevel);
-
-            WeaponStatValues = weaponStatValues.Split(',')
-            .Select((value, index) => new { Index = index + 1, Value = float.Parse(value, CultureInfo.InvariantCulture) })
-            .ToDictionary(x => (WeaponStatType)x.Index, x => x.Value);
-
-            BloodStatValues = bloodStatValues.Split(',')
-            .Select((value, index) => new { Index = index + 1, Value = float.Parse(value, CultureInfo.InvariantCulture) })
-            .ToDictionary(x => (BloodStatType)x.Index, x => x.Value);
-
-            ClassStatSynergies = classStatSynergies
-            .Split(',')
-            .Select((value, index) => new { Value = value, Index = index })
-            .GroupBy(x => x.Index / 3)
-            .ToDictionary(
-                g => (PlayerClass)int.Parse(g.ElementAt(0).Value),
-                g => (
-                    Enumerable.Range(0, g.ElementAt(1).Value.Length / 2)
-                        .Select(j => (WeaponStatType)int.Parse(g.ElementAt(1).Value.Substring(j * 2, 2)))
-                        .ToList(),
-                    Enumerable.Range(0, g.ElementAt(2).Value.Length / 2)
-                        .Select(j => (BloodStatType)int.Parse(g.ElementAt(2).Value.Substring(j * 2, 2)))
+                        .Select(j => (BloodStatType)int.Parse(g.ElementAt(2).Value.Substring(j * 2, 2), CultureInfo.InvariantCulture))
                         .ToList()
                 )
             );
@@ -464,70 +407,9 @@ internal static class DataService
         {
             Core.Log.LogWarning($"Failed to parse config data: {ex}");
         }
-
-        /*
-        try
-        {
-            ConfigData parsedConfigData = new(
-                configData[index++], // prestigeMultiplier
-                configData[index++], // statSynergyMultiplier
-                configData[index++], // maxPlayerLevel
-                configData[index++], // maxLegacyLevel
-                configData[index++], // maxExpertiseLevel
-                configData[index++], // maxFamiliarLevel
-                configData[index++], // maxProfessionLevel no longer used and merits getting cut but that necessitates enough other changes leaving alone for the moment
-                string.Join(",", configData.Skip(index).Take(12)), // Combine the next 11 elements for weaponStatValues
-                string.Join(",", configData.Skip(index += 12).Take(12)), // Combine the following 11 elements for bloodStatValues
-                string.Join(",", configData.Skip(index += 12)) // Combine all remaining elements for classStatSynergies
-            );
-
-            _prestigeStatMultiplier = parsedConfigData.PrestigeStatMultiplier;
-            _classStatMultiplier = parsedConfigData.ClassStatMultiplier;
-
-            _experienceMaxLevel = parsedConfigData.MaxPlayerLevel;
-            _legacyMaxLevel = parsedConfigData.MaxLegacyLevel;
-            _expertiseMaxLevel = parsedConfigData.MaxExpertiseLevel;
-            _familiarMaxLevel = parsedConfigData.MaxFamiliarLevel;
-
-            _weaponStatValues = parsedConfigData.WeaponStatValues;
-            _bloodStatValues = parsedConfigData.BloodStatValues;
-
-            _classStatSynergies = parsedConfigData.ClassStatSynergies;
-
-            return;
-        }
-        catch (Exception e)
-        {
-            Core.Log.LogWarning($"Failed to parse config data second attempt - {e}");
-        }
-        */
     }
     public static void ParsePlayerData(List<string> playerData)
     {
-        /*
-        if (string.IsNullOrEmpty(_version))
-        {
-            if (playerData.Count == 45)
-            {
-                _version = V1_2_2;
-
-                Core.Log.LogInfo($"ECLIPSE[{V1_2_2}]({DateTime.Now}) - {playerData.Count}");
-            }
-
-            if (playerData.Count == 46)
-            {
-
-
-                _version = V1_3;
-                Core.Log.LogInfo($"ECLIPSE[{V1_3}]({DateTime.Now}) - {playerData.Count}");
-            }
-            else
-            {
-                Core.Log.LogWarning("Couldn't parse playerData!");
-            }
-        }
-        */
-
         int index = 0;
 
         ExperienceData experienceData = new(playerData[index++], playerData[index++], playerData[index++], playerData[index++]);
@@ -554,8 +436,6 @@ internal static class DataService
         _expertisePrestige = expertiseData.Prestige;
         _expertiseType = expertiseData.ExpertiseType;
         _expertiseBonusStats = expertiseData.BonusStats;
-
-        // Core.Log.LogWarning($"Expertise bonus stats - {string.Join(", ", _expertiseBonusStats)}");
 
         _familiarProgress = familiarData.Progress;
         _familiarLevel = familiarData.Level;